Refactor input

+++ src/Core/Utility.h
@@ -73,7 +73,7 @@ namespace Math
 	}
 
 	inline float normalizedAngle(float x) {
-		x = fmod(x, 360);
+		x = static_cast<float>(fmod(x, 360));
 		if (x < 0)
 			x += 360;
 		return x;
@@ -81,7 +81,7 @@ namespace Math
 
 	namespace
 	{
-		extern std::default_random_engine generator(time(nullptr));
+		extern std::default_random_engine generator(static_cast<unsigned int>(time(nullptr)));
 	}
 
 	template<typename T>
